Use JP v3 for getStopPlaceDeparturesWithLines, refactor a bit
