+++ src/service.ts
@@ -1,9 +1,6 @@
 import { ApolloClient, InMemoryCache } from '@apollo/client'
 
-import createEnturService, { TransportMode, TransportSubmode } from '@entur/sdk'
-
-import { StopPlaceWithLines, Line } from './types'
-import { unique } from './utils'
+import createEnturClient from '@entur/sdk'
 
 const CLIENT_NAME = process.env.CLIENT_NAME || ''
 
@@ -14,7 +11,7 @@ if (!CLIENT_NAME && process.env.NODE_ENV !== 'production') {
     )
 }
 
-export default createEnturService({
+export default createEnturClient({
     clientName: CLIENT_NAME,
     hosts: {
         journeyPlanner: process.env.JOURNEYPLANNER_HOST_V2,
@@ -30,155 +27,3 @@ export const apolloClient = new ApolloClient({
         'ET-Client-Name': CLIENT_NAME,
     },
 })
-
-function journeyplannerPost<T>(
-    query: string,
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    variables: Record<string, any>,
-    signal: AbortSignal,
-): Promise<T> {
-    return fetch(`${process.env.JOURNEYPLANNER_HOST_V2}/graphql`, {
-        method: 'POST',
-        headers: {
-            'ET-Client-Name': CLIENT_NAME,
-            'Content-Type': 'application/json',
-        },
-        signal,
-        body: JSON.stringify({
-            query,
-            variables,
-        }),
-    }).then((res) => res.json())
-}
-
-interface EstimatedCall {
-    destinationDisplay: {
-        frontText: string
-    }
-    serviceJourney: {
-        line: {
-            id: string
-            transportMode: TransportMode
-            transportSubmode: TransportSubmode
-            publicCode: string
-        }
-        pointsOnLink: {
-            points: string
-        }
-    }
-}
-
-interface StopPlaceWithEstimatedCalls {
-    id: string
-    name: string
-    description: string
-    latitude: number
-    longitude: number
-    transportMode: TransportMode
-    transportSubmode: TransportSubmode
-    estimatedCalls: EstimatedCall[]
-}
-
-function getNumericPublicCode(publicCode: string): number | void {
-    const numberMatch = publicCode.match(/^\d+/)
-    return numberMatch ? Number(numberMatch[0]) : undefined
-}
-
-function publicCodeComparator(a: string, b: string): number {
-    const numA = getNumericPublicCode(a)
-    const numB = getNumericPublicCode(b)
-    if (typeof numA === 'number' && typeof numB === 'number') {
-        return numA - numB
-    }
-    if (typeof numA === 'number') {
-        return -1
-    }
-    if (typeof numB === 'number') {
-        return 1
-    }
-    return a.localeCompare(b, 'no')
-}
-
-function estimatedCallsComparator(a: EstimatedCall, b: EstimatedCall): number {
-    const publicCodeDiff = publicCodeComparator(
-        a.serviceJourney.line.publicCode,
-        b.serviceJourney.line.publicCode,
-    )
-    if (publicCodeDiff !== 0) {
-        return publicCodeDiff
-    }
-
-    return a.destinationDisplay.frontText.localeCompare(
-        b.destinationDisplay.frontText,
-    )
-}
-
-export async function getStopPlacesWithLines(
-    stopPlaceIds: string[],
-    signal: AbortSignal,
-): Promise<StopPlaceWithLines[]> {
-    const variables = { ids: stopPlaceIds }
-    const results = await journeyplannerPost<{
-        data: { stopPlaces: StopPlaceWithEstimatedCalls[] }
-    }>(
-        `query ($ids: [String]!) {
-                stopPlaces(ids: $ids) {
-                    id,
-                    name,
-                    description,
-                    latitude,
-                    longitude,
-                    transportMode,
-                    transportSubmode,
-                    estimatedCalls(numberOfDeparturesPerLineAndDestinationDisplay: 1, timeRange: 604800, numberOfDepartures:200, omitNonBoarding: true) {
-                        destinationDisplay {
-                            frontText
-                        }
-                        serviceJourney {
-                            line {
-                                id
-                                transportMode,
-                                transportSubmode,
-                                publicCode
-                            }
-                            pointsOnLink {
-                                points
-                            }
-                        }
-                    }
-                }
-            }
-            `,
-        variables,
-        signal,
-    )
-
-    const stops: StopPlaceWithLines[] = results.data.stopPlaces.map(
-        (stopPlace) => {
-            try {
-                stopPlace.estimatedCalls
-            } catch {
-                return { ...stopPlace, lines: [] }
-            }
-            const lines = stopPlace.estimatedCalls
-                .sort(estimatedCallsComparator)
-                .map(({ destinationDisplay, serviceJourney }) => ({
-                    ...serviceJourney.line,
-                    name: `${serviceJourney.line.publicCode} ${destinationDisplay.frontText}`,
-                    pointsOnLink: serviceJourney.pointsOnLink.points,
-                }))
-
-            const uniqueLines = unique(
-                lines,
-                (a: Line, b: Line) => a.name === b.name,
-            )
-
-            return {
-                ...stopPlace,
-                lines: uniqueLines,
-            }
-        },
-    )
-
-    return stops
-}
